brand repository modified

+++ src/main/java/com/inventory/deva_inventory/dao/BrandRepository.java
@@ -15,7 +15,7 @@ import org.springframework.stereotype.Repository;
  * @author best
  */
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Long>{
-      public Brand findByBrandId(Long brandId);
+public interface BrandRepository extends JpaRepository<Brand, Integer>{
+      public Brand findByBrandId(Integer brandId);
      public Brand findByBrandName( String brandName);
 }
